+++ hy/models.py
@@ -109,13 +109,13 @@ class HyBytes(HyObject, bytes_type):
 _wrappers[bytes_type] = HyBytes
 
 
-class HySymbol(HyString):
+class HySymbol(HyObject, str_type):
     """
-    Hy Symbol. Basically a String.
+    Hy Symbol. Basically a string.
     """
 
-    def __init__(self, string):
-        self += string
+    def __new__(cls, s=None):
+        return super(HySymbol, cls).__new__(cls, s)
 
 _wrappers[bool] = lambda x: HySymbol("True") if x else HySymbol("False")
 _wrappers[type(None)] = lambda foo: HySymbol("None")
@@ -238,10 +238,10 @@ class HySequence(HyObject, list):
     An abstract type for sequence-like models to inherit from.
     """
 
-    def replace(self, other):
-        for x in self:
-            replace_hy_obj(x, other)
-
+    def replace(self, other, recursive=True):
+        if recursive:
+            for x in self:
+                replace_hy_obj(x, other)
         HyObject.replace(self, other)
         return self